Moved SPACESHIP_SYSTEMS to Planets.py
